pyproject.toml
[tool.poetry]
name = "herd-shepherd-wolf-sim"
version = "0.1.0"
description = "Coevolution simulator with PyTorch + Pygame"
authors = ["Cagnola Giuliana", "Hergenreder Franco", "Saurin Lucas"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
pygame = "^2.6.1"
numpy = "^2.3.4"
matplotlib = "^3.10.7"
torch = "^2.9.0"
torchvision = "^0.24.0"
tqdm = "^4.67.1"
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
ruff = "^0.5.0"
pytest = "^8.0.0"

[tool.poetry.scripts]
run-sim = "scripts.run_simulation:main"
train-agents = "scripts.train_agents:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu116"
priority = "supplemental"
